# backend/swagger.api.yml
swagger: "2.0"
info:
  title: Bitcoin lottery API
  version: "2.0"
host: localhost:5000
basePath: /
schemes:
  - http
paths:
...
  /push/send:
    post:
      tags:
        - notification
      summary: Sends push notifications to the client browser.
      parameters:
        - in: formData
          name: title
          type: string
          description: Title of the notification.
        - in: formData
          name: message
          type: string
          description: Message of the notification.
      responses:
        '200':
          description: OK
  /push/save:
    post:
      tags:
        - notification
      summary: Saves push notifications.
      requestBody:
        content:
          text/html:
            schema:
              type: string
      responses:
        '200':
          description: OK